Use private chains with unlocked accounts for geth
Added in generated account keys that use an empty password so that geth
can unlock those via a js script that is run. This script will unlock
all the accounts and start the mining process so that geth will be able
to accept transactions from truffle. Updated docker testing to use an
image that has the same accounts unlocked at run time and is mining. It
may be worth only allowing tests to be ran through docker.