Decrease gas and allow network warm-up
Slightly decrease the gas that is being sent with transactions. Allow
the development network to almost reach its target gas limit of 1^7.
This allows for the block gas limit to be high enough so that our gas
does not go over the block limit and allows for tests to run. The
downside to this is that it adds 5 minutes to testing because the
network needs to warm up to reach a high block gas limit.