Split APM creation transaction.
To fix gas issues and on deployment and `npm run test`, `newAPM` is
now split into 2 transactions to avoid passing the transaction gas
limit.

On deployment of new APM we must be careful because if only the second
transaction fails the APM would be created but would have wrong
permissions.